RefundDate

+++ src/Controller/Sections_header/RefundController.php
@@ -3,10 +3,13 @@
 namespace App\Controller\Sections_header;
 
 use App\Entity\Invoice;
+use App\Entity\RefundDate;
 use App\Entity\RefundActivity;
+use App\Form\SearchRefundType;
 use App\Form\SearchInvoiceType;
 use App\Entity\IdDetailsManufacturer;
 use App\Repository\InvoiceRepository;
+use App\Repository\RefundDateRepository;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -22,10 +25,11 @@ class RefundController extends AbstractController
         Request $request,
         ValidatorInterface $validator,
         InvoiceRepository $InvoiceRepository,
+        RefundDateRepository $RefundDateRepository,
     ): Response {
 
         /*Подключаем формы */
-        $form_search_refund = $this->createForm(SearchInvoiceType::class);
+        $form_search_refund = $this->createForm(SearchRefundType::class);
 
         /*Валидация формы */
         $form_search_refund->handleRequest($request);
@@ -40,56 +44,59 @@ class RefundController extends AbstractController
             if ($form_search_refund->isValid()) {
                 // dd($request);
                 /* собераем списка по поиску */
-                $number_document_search = $request->request->all()['search_invoice']['search_number_document'];
-                $s_data_invoice_search = $request->request->all()['search_invoice']['s_data_invoice'];
-                $po_data_invoice_search = $request->request->all()['search_invoice']['po_data_invoice'];
-                $id_counterparty_search = $request->request->all()['search_invoice']['search_id_counterparty'];
-                $part_numbers_search = strtolower($request->request->all()['search_invoice']['search_id_details']);
-                $manufacturers_search = strtolower($request->request->all()['search_invoice']['search_id_manufacturer']);
-                $s_price_search = $request->request->all()['search_invoice']['s_price'];
-                $po_price_search = $request->request->all()['search_invoice']['po_price'];
-
+                $number_document_search = $request->request->all()['search_refund']['search_number_document_refund'];
+                $s_data_invoice_search = $request->request->all()['search_refund']['s_data_invoice_refund'];
+                $po_data_invoice_search = $request->request->all()['search_refund']['po_data_invoice_refund'];
+                $id_counterparty_search = $request->request->all()['search_refund']['search_id_counterparty_refund'];
+                $part_numbers_search = strtolower($request->request->all()['search_refund']['search_id_details_refund']);
+                $manufacturers_search = strtolower($request->request->all()['search_refund']['search_id_manufacturer_refund']);
+                $s_price_search = $request->request->all()['search_refund']['s_price_refund'];
+                $po_price_search = $request->request->all()['search_refund']['po_price_refund'];
+                $data_refund = $request->request->all()['search_refund']['data_refund'];
 
                 if ($number_document_search) {
 
-                    $arr_refund[] = $doctrine->getRepository(Invoice::class)
-                        ->findBy(['number_document' => $number_document_search]);
+                    $arr_refund[] = $RefundDateRepository->findByNumberDocument($number_document_search);
                 } elseif ($s_data_invoice_search && $po_data_invoice_search) {
 
-                    $arr_refund[] = $InvoiceRepository
-                        ->findByDate($s_data_invoice_search, $po_data_invoice_search);
+                    $arr_refund[] = $RefundDateRepository
+                        ->findByDateInvoice($s_data_invoice_search, $po_data_invoice_search);
                 } elseif ($s_data_invoice_search && !$po_data_invoice_search) {
 
-                    $arr_refund[] = $InvoiceRepository
-                        ->findBySDate($s_data_invoice_search);
+                    $arr_refund[] = $RefundDateRepository
+                        ->findBySDateInvoice($s_data_invoice_search);
                 } elseif (!$s_data_invoice_search && $po_data_invoice_search) {
 
-                    $arr_refund[] = $InvoiceRepository
-                        ->findByPoDate($po_data_invoice_search);
+                    $arr_refund[] = $RefundDateRepository
+                        ->findByPoDateInvoice($po_data_invoice_search);
                 } elseif ($id_counterparty_search) {
 
-                    $arr_refund[] = $doctrine->getRepository(Invoice::class)
-                        ->findBy(['id_counterparty' => $id_counterparty_search]);
+                    $arr_refund[] = $RefundDateRepository
+                        ->findByCounterparty($id_counterparty_search);
                 } elseif ($part_numbers_search) {
 
-                    $arr_refund[] = $InvoiceRepository
-                        ->findBySearchNumber($part_numbers_search);
+                    $arr_refund[] = $RefundDateRepository
+                        ->findByPartNumber($part_numbers_search);
                 } elseif ($manufacturers_search) {
 
-                    $arr_refund[] = $InvoiceRepository
-                        ->findBySearchManufacturers($manufacturers_search);
+                    $arr_refund[] = $RefundDateRepository
+                        ->findByPartManufacturers($manufacturers_search);
                 } elseif ($s_price_search && $po_price_search) {
 
-                    $arr_refund[] = $InvoiceRepository
-                        ->findByPrice($s_price_search, $po_price_search);
+                    $arr_refund[] = $RefundDateRepository
+                        ->findByPriceRefund($s_price_search, $po_price_search);
                 } elseif ($s_price_search && !$po_price_search) {
 
-                    $arr_refund[] = $InvoiceRepository
-                        ->findBySPrice($s_price_search);
+                    $arr_refund[] = $RefundDateRepository
+                        ->findBySPriceRefund($s_price_search);
                 } elseif (!$s_price_search && $po_price_search) {
 
-                    $arr_refund[] = $InvoiceRepository
-                        ->findByPoPrice($po_price_search);
+                    $arr_refund[] = $RefundDateRepository
+                        ->findByPoPriceRefund($po_price_search);
+                } elseif ($data_refund) {
+
+                    $arr_refund[] = $RefundDateRepository
+                        ->findByDataRefund($data_refund);
                 }
             } else {
 
@@ -119,11 +126,11 @@ class RefundController extends AbstractController
                 return $this->redirectToRoute('refund');
             }
         } else {
-            $arr_refund[] = $doctrine->getRepository(Invoice::class)->findAll();
-            $arr_refund[] = $InvoiceRepository->findByRefund();
+
+            $arr_refund[] = $RefundDateRepository->findByRefund();
         }
 
-        dd($arr_refund);
+        //dd($arr_refund);
         return $this->render('refund/refund.html.twig', [
             'title_logo' => 'Возвращенные детали',
             'legend_search' => 'Поиск',
@@ -143,13 +150,21 @@ class RefundController extends AbstractController
     #[Route('/edit_refund', name: 'edit_refund')]
     public function EditRefund(ManagerRegistry $doctrine, Request $request): Response
     {
-        $id = $request->request->all()['edit_refund'];
+        if ($request->request->all()) {
+            $id = $request->request->all()['edit_refund'];
 
-        $edit_refund = $doctrine->getRepository(Invoice::class)->find($id);
+            $edit_refund = $doctrine->getRepository(Invoice::class)->find($id);
 
-        $edit_refund->setRefund(1);
+            $edit_refund->setRefund(1);
 
-        $doctrine->getManager()->flush();
+            $delete_refund_date = $doctrine->getRepository(RefundDate::class)
+                ->findOneBy(['id_invoice_refund_date' => $id]);
+
+            $entityManager = $doctrine->getManager();
+            $entityManager->remove($delete_refund_date);
+            $entityManager->flush();
+        }
+
 
         return $this->redirectToRoute('refund');
     }
@@ -158,14 +173,16 @@ class RefundController extends AbstractController
     #[Route('/delete_refund', name: 'delete_refund')]
     public function deleteRefund(ManagerRegistry $doctrine, Request $request): Response
     {
+        // dd($request->request->all());
+        if ($request->request->all()) {
+            $id_delete_refund = $request->request->all()['delete_refund'];
 
-        $id_delete_refund = $request->request->all()['delete_refund'];
+            $delete_refund = $doctrine->getRepository(Invoice::class)->find($id_delete_refund);
 
-        $delete_refund = $doctrine->getRepository(Invoice::class)->find($id_delete_refund);
-
-        $entityManager = $doctrine->getManager();
-        $entityManager->remove($delete_refund);
-        $entityManager->flush();
+            $entityManager = $doctrine->getManager();
+            $entityManager->remove($delete_refund);
+            $entityManager->flush();
+        }
 
         return $this->redirectToRoute('refund');
     }
@@ -174,25 +191,27 @@ class RefundController extends AbstractController
     #[Route('/edit_activity', name: 'edit_activity')]
     public function EditActivity(ManagerRegistry $doctrine, Request $request): Response
     {
-        $id_activity = '';
+        if ($request->request->all()) {
 
-        $id = $request->request->all()['edit_activity'];
+            $id = $request->request->all()['edit_activity'];
 
-        $edit_activity = $doctrine->getRepository(Invoice::class)->find($id);
+            $edit_activity = $doctrine->getRepository(Invoice::class)->find($id);
 
-        $refund_activity = $edit_activity->getIdRefundActivity()->getId();
+            $refund_activity = $edit_activity->getIdRefundActivity()->getId();
 
-        if ($refund_activity == 1) {
-            $id_activity = 2;
-        } elseif ($refund_activity == 2) {
-            $id_activity = 1;
+            $id_activity = '';
+            if ($refund_activity == 1) {
+                $id_activity = 2;
+            } elseif ($refund_activity == 2) {
+                $id_activity = 1;
+            }
+
+            $id_refund_activit = $doctrine->getRepository(RefundActivity::class)->find($id_activity);
+
+            $edit_activity->setIdRefundActivity($id_refund_activit);
+
+            $doctrine->getManager()->flush();
         }
-
-        $id_refund_activit = $doctrine->getRepository(RefundActivity::class)->find($id_activity);
-
-        $edit_activity->setIdRefundActivity($id_refund_activit);
-
-        $doctrine->getManager()->flush();
 
         return $this->redirectToRoute('refund');
     }
